Changed spins and weights to be integer based instead of double

+++ NetlistPartitioning/Spin.cpp
@@ -10,7 +10,7 @@ Spin::Spin() {
     next_spin = 1;
 }
 
-void Spin::add_to_graph(double* _weights, Spin* _adjacent, int _neighbours) {
+void Spin::add_to_graph(int* _weights, Spin* _adjacent, int _neighbours) {
     adjacent = _adjacent;
     weights = _weights;
     neighbours = _neighbours;
@@ -20,8 +20,8 @@ void Spin::randomize() {
     spin = randf() > 0.5 ? -1 : 1;
 }
 
-double Spin::energy_adjacent() {
-    double total = 0;
+int Spin::energy_adjacent() {
+    int total = 0;
     for (int i = 0; i < neighbours; i++) {
         total += spin * adjacent[i].spin * weights[i];
     }
