+++ NetlistPartitioning/Spin.h
@@ -3,16 +3,16 @@
 
 class Spin {
 public:
-    double spin, next_spin;
+    int spin, next_spin;
     Spin* adjacent;
-    double* weights;
+    int* weights;
     int neighbours;
     std::string name;
 
     Spin();
-    void add_to_graph(double*, Spin*, int);
+    void add_to_graph(int*, Spin*, int);
     void flip();
-    double energy_adjacent();
+    int energy_adjacent();
     void next_state(bool flip);
     void randomize();
 };
